Fix DM breaking bot and add comments

+++ index.js
@@ -45,16 +45,18 @@ client.on('ready', () => {
 });
 
 client.on('message', message => {
+    if(!message.guild || !message.guild.available) return; // Stop if message is DM or server is offline
+
     let index = indexOfObjectByName(client.servers, message.guild.name);
     var server = client.servers[index];
 
-    if(message.type === "PINS_ADD") return message.delete();
-    if (!message.content.startsWith(server.prefix) || message.author.bot) return;
+    if(message.type === "PINS_ADD") return message.delete(); // Delete 'user' pinned a message' message
+    if (!message.content.startsWith(server.prefix) || message.author.bot) return; // Stop if message does not start with command prefix
 
-    const args = message.content.slice(server.prefix.length).split(/ +/);
-    const commandName = args.shift().toLowerCase();
+    const args = message.content.slice(server.prefix.length).split(/ +/); // remove command prefix from message
+    const commandName = args.shift().toLowerCase(); // extract command used from message
 
-    if (index == -1) {
+    if (index == -1) { // if server not already included in channels.json
         client.servers.push({
             name: message.guild.name,
             lastPrefix: "!", prefix: "~",
@@ -64,34 +66,34 @@ client.on('message', message => {
         index = client.servers.length - 1;
     }
 
-    const command = client.commands.get(commandName) || client.commands.find(cmd => cmd.aliases && cmd.aliases.includes(commandName));
+    const command = client.commands.get(commandName) || client.commands.find(cmd => cmd.aliases && cmd.aliases.includes(commandName)); // find the command used
 
     if (!command) return;
 
     let permissions = ['user'];
 
-    if (message.member.roles.some(val => server.role === val)) {
+    if (message.member.roles.some(val => server.role === val)) { // if bot owner, give admin role
         permissions.push('admin');
     }
 
-    if (message.guild.owner == message.member) {
+    if (message.guild.owner == message.member) { // if server owner, give admin and owner role
         permissions.push('admin');
         permissions.push('owner');
     }
 
-    if (message.member.id == "83002230331932672") {
+    if (message.member.id == "140209881280937984") { // if CheeseLiker, give all permissions
         permissions.push('admin');
         permissions.push('owner');
         permissions.push('botowner');
     }
 
-    if (command.permission && !(permissions.indexOf(command.permission) > -1)) {
+    if (command.permission && !(permissions.indexOf(command.permission) > -1)) { // user does not have command permisions
         // user not allowed...
         message.reply(`You do not have the required permission \`${command.permission}\` to run this command.`);
         return;
     }
 
-    if (command.args && !args.length) {
+    if (command.args && !args.length) { // if needed arguments not provided
         let reply = `You didn't provide any arguments, ${message.author}!`;
         if (command.usage) {
             reply += `\nThe proper usage would be: \`${server.prefix}${command.name} ${command.usage}\``;
@@ -99,22 +101,23 @@ client.on('message', message => {
         return message.channel.send(reply);
     }
 
-    try {
+    try { // Run the command
         command.execute(client, message, args);
     }
     catch (error) {
         logger.error(error);
-        message.reply('there was an error trying to execute that command!');
+        message.reply('There was an error trying to execute that command!');
     }
 });
 
 client.on("guildMemberAdd", member => {
-    member.addRole(['476662770209783808','484332514110996490'])
-        .then(console.log)
-        .catch(console.error);
+    member.addRole(['476662770209783808','484332514110996490']) // Add roles to user on join
+        .then(() => logger.info("Added Roles to New User: " + member.name))
+        .catch((e) => logger.error("Add Roles Error: " + e));
 });
 
-function indexOfObjectByName(array, value) {
+function indexOfObjectByName(array, value) { 
+    // Helper to get Object's Index in an array using it's name
     for (let i = 0; i < array.length; i++) {
         if (array[i].name.toLowerCase().trim() === value.toLowerCase().trim()) {
             return i;
@@ -124,6 +127,7 @@ function indexOfObjectByName(array, value) {
 }
 
 function tick() {
+    // Get Twitch Channel info from API every tick
     client.servers.forEach((server) => {
         try {
             client.twitchapi.users.usersByName({ users: server.twitchChannels.map(x => x.name) }, getChannelInfo.bind(this, server))
@@ -136,6 +140,7 @@ function tick() {
 }
 
 function getChannelInfo(server, err, res) {
+    // Handle the Twitch API response for each Twitch Channel
     if (!res) return;
     if (err) logger.error(`Error in getChannelInfo: ${err}`);
 
@@ -167,19 +172,20 @@ function createEmbed(server, twitchChannel, res) {
 
 
 function postDiscord(server, twitchChannel, err, res) {
+    // Check if Twitch Channel is Live and Make/Update the Live announcement in Discord
     if (!res) return;
     if (err) logger.error(`Error in postDiscord: ${err}`);
-    if (server.discordChannels.length == 0) return;
+    if (server.discordChannels.length == 0) return; // stop if no Discord Channel set to post in
 
     if (res.stream != null && twitchChannel.messageid == null) {
-        // Do new message code
+        // Do new Discord message code
         try {
             const guild = client.guilds.find(guild => guild.name === server.name);
             const discordChannel = guild.channels.find(discordChannel => discordChannel.name === server.discordChannels[0]);
             const discordEmbed = createEmbed(server, twitchChannel, res);
-            const discordLive = res.stream.channel.display_name + " is <@&484332514110996490>!"
+            const discordPing = res.stream.channel.display_name + " is <@&484332514110996490>!" // message with @mention used before embed
 
-            discordChannel.send(discordLive, { embed: discordEmbed }).then(
+            discordChannel.send(discordPing, { embed: discordEmbed }).then( // send the Discord Live Message
                 (message) => {
                     logger.info(`[${server.name}/${discordChannel.name}] Now Live: ${twitchChannel.name}`)
                     twitchChannel.messageid = message.id
@@ -192,14 +198,14 @@ function postDiscord(server, twitchChannel, err, res) {
             logger.error(`Error in postDiscord new msg: ${err}`);
         }
     } else if (res.stream != null && twitchChannel.messageid != null) {
-        // Do edit message code
+        // Do edit Discord message code
         try {
             const guild = client.guilds.find(guild => guild.name === server.name);
             const discordChannel = guild.channels.find(discordChannel => discordChannel.name === server.discordChannels[0]);
             const discordEmbed = createEmbed(server, twitchChannel, res);
             const discordLive = res.stream.channel.display_name + " is <@&484332514110996490>!"
 
-            discordChannel.fetchMessage(twitchChannel.messageid).then(
+            discordChannel.fetchMessage(twitchChannel.messageid).then( // edit/update the message
                 message => message.edit(discordLive, { embed: discordEmbed }).then((message) => {
                     logger.info(`[${server.name}/${discordChannel.name}] Channel Update: ${twitchChannel.name}`)
                     twitchChannel.messageid = message.id
@@ -211,13 +217,12 @@ function postDiscord(server, twitchChannel, err, res) {
             logger.error(`Error in postDiscord edit msg: ${err}`);
         }
     } else if (res.stream == null && twitchChannel.messageid != null) {
-        // Do delete message code
+        // Do delete Discord message code
         try {
             const guild = client.guilds.find(guild => guild.name === server.name);
             const discordChannel = guild.channels.find(discordChannel => discordChannel.name === server.discordChannels[0]);
-            // const discordEmbed = createEmbed(server, twitchChannel, res);
 
-            discordChannel.fetchMessage(twitchChannel.messageid).then(
+            discordChannel.fetchMessage(twitchChannel.messageid).then( // delete the message once offline
                 message => message.delete().then((message) => {
                     logger.info(`[${server.name}/${discordChannel.name}] Channel Offline: ${twitchChannel.name}`)
                     twitchChannel.messageid = null
@@ -231,6 +236,7 @@ function postDiscord(server, twitchChannel, err, res) {
 }
 
 function savechannels() {
+    // Save Discord / Twitch Channel Information (client.servers) to channels.json
     logger.info("Saving channels to " + channelPath);
     fs.writeFileSync(channelPath, JSON.stringify(client.servers, null, 4));
     logger.info("Done");
@@ -253,7 +259,7 @@ process.on("SIGINT", exitHandler.bind(null, { exit: true }));
 process.on("SIGTERM", exitHandler.bind(null, { exit: true }));
 process.on("uncaughtException", exitHandler.bind(null, { exit: true }));
 
-try {
+try { // Login to discord client
     client.login(discordtoken)
 } catch (err) {
     logger.error(`Error in Discord login: ${err}`);
